# docs/swagger.yaml
basePath: /api/v1
definitions:
  apitypes.MassCalculationJSON:
    properties:
      creator_login:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_form:
        type: string
      id:
        type: integer
      moderator_login:
        type: string
      output_koef:
        type: number
      status:
        type: string
    type: object
...
  apitypes.ReactionJSON:
    properties:
      conversation_factor:
        type: number
      description:
        type: string
      id:
        type: integer
      img_link:
        type: string
      product:
        type: string
      reagent:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для работы с реакциями
  title: Reactions API
  version: "1.0"
paths:
...
  /reactions:
    get:
      description: Возвращает все реакции или фильтрует по параметру reaction_title.
      parameters:
      - description: Фильтр по заголовку реакции
        in: query
        name: reaction_title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список реакций
          schema:
            items:
              $ref: '#/definitions/apitypes.ReactionJSON'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список реакций
      tags:
      - reactions
    post:
      consumes:
      - application/json
      description: Создаёт новую реакцию, возвращает Location в заголовках и объект
        реакции.
      parameters:
      - description: Данные реакции
        in: body
        name: reaction
        required: true
        schema:
          $ref: '#/definitions/apitypes.ReactionJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Реакция создана
          schema:
            $ref: '#/definitions/apitypes.ReactionJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать реакцию
      tags:
      - reactions
  /reactions/{id}:
    delete:
      description: Логическое удаление реакции. Возвращает {"status":"deleted"}.
      parameters:
      - description: ID реакции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Реакция не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить реакцию
      tags:
      - reactions
    get:
      description: Возвращает реакцию по id.
      parameters:
      - description: ID реакции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Реакция
          schema:
            $ref: '#/definitions/apitypes.ReactionJSON'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Реакция не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить реакцию по ID
      tags:
      - reactions
    put:
      consumes:
      - application/json
      description: Обновляет реакцию по ID.
      parameters:
      - description: ID реакции
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные реакции
        in: body
        name: reaction
        required: true
        schema:
          $ref: '#/definitions/apitypes.ReactionJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённая реакция
          schema:
            $ref: '#/definitions/apitypes.ReactionJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Реакция не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить реакцию
      tags:
      - reactions
  /reactions/{id}/add-to-calculation:
    post:
      description: Добавляет реакцию в черновик пользователя (если черновика нет —
        создаёт). Возвращает сам расчет.
      parameters:
      - description: ID реакции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Если добавление в существующий черновик
          schema:
            $ref: '#/definitions/apitypes.MassCalculationJSON'
        "201":
          description: Если был создан новый черновик
          schema:
            $ref: '#/definitions/apitypes.MassCalculationJSON'
        "400":
          description: Ошибка запроса или авторизации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Реакция/калькуляция не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить реакцию в текущий черновик расчета
      tags:
      - mass-calculations
  /reactions/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: 'Загружает файл изображения и возвращает объект вида {"status":"uploaded",
        "reaction": <ReactionJSON>}.'
      parameters:
      - description: ID реакции
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: status (string) и reaction (apitypes.ReactionJSON)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос/файл
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Реакция не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение для реакции
      tags:
      - reactions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"